Renamed tests so they're auto-discovered by NoseTest Fixed error in TestTeam Fixed error in TestStringQuestion
